Update README.md
his project includes solving a real-world business problem of sales forecasting and building up a machine learning model for the sales prediction. Our main goal here is to predict the sales for six weeks for each store and find out the factors influencing in it and recommend ways in order to improve the numbers.